refactor(navigation): split specialized definition helpers (#0000)
